fix: correct integer shift and random device initialization
- Fix potential overflow in item price calculation by using 64-bit shift
- Initialize random device properly to avoid undefined behavior
- Update Visual Studio version and platform toolset to v143
- Add security warning disables for CRT and Winsock functions

+++ GameServer/GameServer/GMHolyItem.cpp
@@ -196,7 +196,7 @@ void CGMHolyItem::GCInventoryListSend(int aIndex)
 
 	PMSG_INVENTORY_HOLY info;
 
-	for (size_t n = 0; n < HOLY_INVENTORY_SIZE; n++)
+	for (BYTE n = 0; n < HOLY_INVENTORY_SIZE; n++)
 	{
 		if (lpObj->HolyInventory[n].bItemindex != -1)
 		{
